+++ src/utils.js
@@ -887,4 +887,28 @@ export const clearStorage = () => {
   });
 };
 
+export const setRecordForDeletion = async (okapi, id, tenantId) => {
+  const {
+    url,
+    token,
+  } = okapi;
+  const path = `${url}/inventory/instances/${id}/mark-deleted`;
+
+  const response = await fetch(path, {
+    method: 'DELETE',
+    headers: {
+      [OKAPI_TENANT_HEADER]: tenantId,
+      [CONTENT_TYPE_HEADER]: 'application/json',
+      ...(token && { [OKAPI_TOKEN_HEADER]: token }),
+    },
+    credentials: 'include',
+  });
+
+  if (!response.ok) {
+    throw response;
+  }
+
+  return response;
+};
+
 export const parseEmptyFormValue = value => value;