+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## [10.1.0] (IN PROGRESS)
 
+* Remove unused code related to auto-open record detail view. Refs UIIN-2819.
+
 ## [10.0.1] (IN PROGRESS)
 
 * Keep query and results list when switching Browse options. Refs UIIN-2802.
+* Set central tenant id in the request when Member tenant deletes a shared record. Fixes UIIN-2784.
 * Apply staff suppress filter for first search in Holdings/Items. Fixes UIIN-2814.
 
 ## [11.0.0](https://github.com/folio-org/ui-inventory/tree/v11.0.0) (2024-03-21)